UI and Tutorial changes
This is a major update to the UI, data logic, and tutorial process.

On first visit (as determined by checking localStorage), the user is walked through a series of prompts before beginning the prompts for the day. This requires creating two new XML responses, help_end and help. The prior messy logic has been updated so that the browser must choose to get *all* variables from localStorage, or else use a newly input data object.

The UI changes include changing the font to Schibsted Grotesk, varying the grid heights, and updating transition and animation logic. Debug statements have been set up for all POST requests to / as well.